refactor: Break the Table file into multiple files
